backend/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.11
aiosignal==1.3.2
aiosqlite==0.17.0
annotated-types==0.7.0
anyio==4.9.0
asyncpraw==7.8.1
asyncprawcore==2.4.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
curl_cffi==0.11.2
distro==1.9.0
fastapi==0.115.12
frozendict==2.4.6
frozenlist==1.6.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
lxml==5.4.0
multidict==6.4.4
multitasking==0.0.11
numpy==2.3.0
openai==1.84.0
pandas==2.3.0
peewee==3.18.1
platformdirs==4.3.8
praw==7.8.1
prawcore==2.4.0
propcache==0.3.1
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytrends==4.9.2
pytz==2025.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
update-checker==0.18.0
urllib3==2.4.0
uvicorn==0.34.3
websocket-client==1.8.0
websockets==15.0.1
yarl==1.20.0
yfinance==0.2.62
